+++ qa/Vagrantfile
@@ -17,6 +17,10 @@ Vagrant.configure(2) do |config|
         sh.path = "sys/#{platform.type}/bootstrap.sh"
         sh.privileged = true
       end
+      machine.vm.provision :shell do |sh|
+        sh.path = "sys/#{platform.type}/user_bootstrap.sh"
+        sh.privileged = false
+      end
     end
   end
 end
