--- qa/platforms.json
@@ -1,13 +1,0 @@
-{ 
-  "ubuntu-1204": { "box": "elastic/ubuntu-12.04-x86_64", "type": "debian" },
-  "ubuntu-1404": { "box": "elastic/ubuntu-14.04-x86_64", "type": "debian" },
-  "ubuntu-1504": { "box": "elastic/ubuntu-15.04-x86_64", "type": "debian" },
-  "centos-6": { "box": "elastic/centos-6-x86_64", "type": "redhat" },
-  "centos-7": { "box": "elastic/centos-7-x86_64", "type": "redhat" },
-  "oel-6": { "box": "elastic/oraclelinux-6-x86_64", "type": "redhat" },
-  "oel-7": { "box": "elastic/oraclelinux-7-x86_64", "type": "redhat" },
-  "fedora-22": { "box": "elastic/fedora-22-x86_64", "type": "redhat" },
-  "debian-8": { "box": "elastic/debian-8-x86_64", "type": "debian" },
-  "opensuse-13": { "box": "elastic/opensuse-13-x86_64", "type": "suse" },
-  "sles-12": { "box": "elastic/sles-12-x86_64", "type": "suse" }
-}
